requirements.txt
python-telegram-bot==20.7
aiosqlite==0.19.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
asyncio-throttle==1.0.2
schedule==1.2.0
pytz==2023.3
requests==2.31.0 